backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-mail>=0.10.0",
    "gspread>=6.2.1",
    "gunicorn>=23.0.0",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.2.1",
    "werkzeug>=3.1.4",
    "firebase-admin>=6.2.0",
    "supabase-py>=0.4.0",
    "stripe>=3.23.0",
    "paystack>=0.2.0",
    "flask-bcrypt>=1.0.1"
]
